+++ packages/systems/tests/main.test.ts
@@ -9,7 +9,6 @@ test("Find by name", () => {
     effect: null,
     id: 30000142,
     name: "Jita",
-    regionId: 10000002,
     region: {
       id: 10000002,
       name: "The Forge",
@@ -82,7 +81,6 @@ test("Find by ID", () => {
         },
       ],
     },
-    regionId: 11000029,
     region: {
       id: 11000029,
       name: "E-R00029",