+++ packages/systems/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added classes 14-18 to `System.whClass` type.
 - Removed class 25 (Pochven) from `System.whClass` type as Pochven is not a wormhole region and Pochven systems can be identified using `System.region`.
+- Removed deprecated `regionId` property as duplicate of `region.id`.
 
 ### Additions